refactor: Extract flow events to own file
This solves a circular dependency problem between budget-insight
and ConnectionFlow
