+++ packages/cozy-harvest-lib/src/models/ConnectionFlow.js
@@ -24,32 +24,27 @@ import assert from '../assert'
 import * as accounts from '../helpers/accounts'
 import * as triggersModel from '../helpers/triggers'
 import sentryHub from '../sentry'
+import {
+  ERRORED,
+  ERROR_EVENT,
+  LOGIN_SUCCESS,
+  LOGIN_SUCCESS_EVENT,
+  SUCCESS,
+  SUCCESS_EVENT,
+  IDLE,
+  UPDATE_EVENT,
+  RUNNING_TWOFA,
+  TWO_FA_REQUEST,
+  TWO_FA_REQUEST_EVENT,
+  TWO_FA_MISMATCH,
+  TWO_FA_MISMATCH_EVENT,
+  WAITING_TWOFA,
+  CREATING_ACCOUNT,
+  PENDING,
+  JOB_EVENTS
+} from './flowEvents'
 
 const JOBS_DOCTYPE = 'io.cozy.jobs'
-
-// events
-export const ERROR_EVENT = 'error'
-export const LOGIN_SUCCESS_EVENT = 'loginSuccess'
-export const SUCCESS_EVENT = 'success'
-export const TRIGGER_LAUNCH_EVENT = 'TRIGGER_LAUNCH_EVENT'
-
-export const TWO_FA_REQUEST_EVENT = 'twoFARequest'
-export const TWO_FA_MISMATCH_EVENT = 'twoFAMismatch'
-export const UPDATE_EVENT = 'update'
-
-// statuses
-export const IDLE = 'IDLE'
-export const CREATING_ACCOUNT = 'CREATING_ACCOUNT'
-export const PENDING = 'PENDING'
-export const ERRORED = 'ERRORED'
-export const LOGIN_SUCCESS = 'LOGIN_SUCCESS'
-export const SUCCESS = 'SUCCESS'
-export const RUNNING_TWOFA = 'RUNNING_TWOFA'
-export const WAITING_TWOFA = 'WAITING_TWOFA'
-export const TWO_FA_MISMATCH = 'TWO_FA_MISMATCH'
-export const TWO_FA_REQUEST = 'TWO_FA_REQUEST'
-
-const JOB_EVENTS = [ERROR_EVENT, LOGIN_SUCCESS_EVENT, SUCCESS_EVENT]
 
 const eventToStatus = {
   [ERROR_EVENT]: ERRORED,
@@ -577,6 +572,7 @@ export class ConnectionFlow {
   }
 
   getKonnectorPolicy() {
+    console.log('gkp')
     return findKonnectorPolicy(this.konnector)
   }
 
@@ -586,6 +582,8 @@ export class ConnectionFlow {
       throw new Error(
         'A konnector policy is needed to be able to detect additional information'
       )
+    } else if (!konnectorPolicy.getAdditionalInformationNeeded) {
+      throw new Error('No getAdditionalInformationNeeded in konnector policy')
     } else {
       const fields = konnectorPolicy.getAdditionalInformationNeeded(this)
       return fields
