+++ packages/cozy-harvest-lib/src/models/flowEvents.js
@@ -1,0 +1,23 @@
+// events
+export const ERROR_EVENT = 'error'
+export const LOGIN_SUCCESS_EVENT = 'loginSuccess'
+export const SUCCESS_EVENT = 'success'
+export const TRIGGER_LAUNCH_EVENT = 'TRIGGER_LAUNCH_EVENT'
+
+export const TWO_FA_REQUEST_EVENT = 'twoFARequest'
+export const TWO_FA_MISMATCH_EVENT = 'twoFAMismatch'
+export const UPDATE_EVENT = 'update'
+
+// statuses
+export const IDLE = 'IDLE'
+export const CREATING_ACCOUNT = 'CREATING_ACCOUNT'
+export const PENDING = 'PENDING'
+export const ERRORED = 'ERRORED'
+export const LOGIN_SUCCESS = 'LOGIN_SUCCESS'
+export const SUCCESS = 'SUCCESS'
+export const RUNNING_TWOFA = 'RUNNING_TWOFA'
+export const WAITING_TWOFA = 'WAITING_TWOFA'
+export const TWO_FA_MISMATCH = 'TWO_FA_MISMATCH'
+export const TWO_FA_REQUEST = 'TWO_FA_REQUEST'
+
+export const JOB_EVENTS = [ERROR_EVENT, LOGIN_SUCCESS_EVENT, SUCCESS_EVENT]
